requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
click==8.1.3
colorama==0.4.6
discord==2.2.2
discord.py==2.2.2
Fenv==0.0.12.2
filelock==3.12.0
frozenlist==1.3.3
h11==0.14.0
multidict==6.0.4
mypy-extensions==1.0.0
newsapi-python==0.2.7
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pypresence==4.2.1
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
six==1.16.0
soupsieve==2.4.1
tomli==2.0.1
tzdata==2023.3
tzlocal==4.3
virtualenv==20.22.0
yarl==1.9.1
